+++ src/CodeMap/MixedCaseMap.php
@@ -40,4 +40,12 @@ final class MixedCaseMap implements CodeMap
         }
         return $codeMap;
     }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function getLengths(): array
+    {
+        return [1];
+    }
 }