+++ src/CodeMap/UpperCaseMap.php
@@ -33,4 +33,11 @@ final class UpperCaseMap implements CodeMap
         }
         return $codeMap;
     }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function getLengths(): array
+    {
+        return [1];
 }